fix(mangastream): parse ints as u64 instead of u32
Garbage was parsed as a u32, which could not handle large numbers

+++ src/rust/mangastream/template/src/helper.rs
@@ -312,7 +312,7 @@ pub fn get_permanet_url(original_url: String) -> String {
 
 	// check if the garbage is a 10 digit number to prevent removing the wrong part
 	// of the url the garbage should always be a 10 digit number
-	if garbage.parse::<u32>().is_ok() && garbage.len() == 10 {
+	if garbage.parse::<u64>().is_ok() && garbage.len() == 10 {
 		// remove the garbage from the url
 		// example https://luminousscans.com/series/1671729411-a-bad-person/
 		// will return https://luminousscans.com/series/a-bad-person
